Large amount of additions
Fixed a few problems in the previous files, and filled out much more of
the implementation. I have some ideas on how it will all come together,
but will need to get with Roger to see how io, maps, and markup will
work, and with Ethan to get item and creature functionality down. Until
I have an interface for Creatures, I can't implement a battle system.